agent-tui: Move App logic out of main
This will allow for minimal code duplication while agent-tui is built in
assisted-installer-agent as well.
