+++ tools/agent_tui/app.go
@@ -1,0 +1,24 @@
+package agent_tui
+
+import (
+	"github.com/openshift/agent-installer-utils/tools/agent_tui/dialogs"
+	"github.com/openshift/agent-installer-utils/tools/agent_tui/forms"
+	"github.com/openshift/agent-installer-utils/tools/agent_tui/newt"
+	"github.com/rivo/tview"
+)
+
+func App() {
+	app := tview.NewApplication()
+	pages := tview.NewPages()
+
+	background := tview.NewBox().
+		SetBorder(false).
+		SetBackgroundColor(newt.ColorBlue)
+
+	pages.AddPage("background", background, true, true).
+		AddPage("Node0", forms.IsNode0Modal(app, pages), true, true)
+
+	if err := app.SetRoot(pages, true).Run(); err != nil {
+		dialogs.PanicDialog(app, err)
+	}
+}
